requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Caching==2.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0.0
Faker==24.4.0
google-genai
